+++ package.json
@@ -1,4 +1,5 @@
 {
+  
   "dependencies": {
     "three": "^0.137.5"
   }
