--- style.css
@@ -1,15 +1,0 @@
-body { 
-    margin: 0;
-    height: 100vh;
-}
-canvas { 
-    display: block;
-}
-
-#wallCoord {
-    padding: 20px;
-}
-
-.addRoomDiv {
-    display: none;
-}
